Update ZugferdProfileResolver.php
Fixes #133

The new version of php_zugferd makes use of libxml_get_last_error() in order to verify that SimpleXMLElement() works.

However, if the user code (the code that calls php_zugferd) had accidentally raised an libxml error, then the error status will be preserved until php_zugferd calls libxml_get_last_error(). In other words: The user code had an error, completely unrelated to ZUGFeRD, and ZUGFeRD raises an error telling that the XML is wrong, which is not true.

This PR fixes the issue by making sure that the libxml error storage is clean before checking the SimpleXML result.